move solve into ir
