+++ src/coherence.rs
@@ -3,7 +3,7 @@ use petgraph::prelude::*;
 use errors::Result;
 use ir::{self, ItemId};
 use rust_ir::Program;
-use solve::SolverChoice;
+use ir::solve::SolverChoice;
 use std::sync::Arc;
 
 mod solve;
