+++ src/ir/solve/infer/ucanonicalize.rs
@@ -57,7 +57,7 @@ crate struct UCanonicalized<T> {
     crate quantified: UCanonical<T>,
 
     /// A map between the universes in `quantified` and the original universes
-    pub(in solve) universes: UniverseMap,
+    pub(in ir::solve) universes: UniverseMap,
 }
 
 /// Maps the universes found in the `u_canonicalize` result (the
@@ -66,7 +66,7 @@ crate struct UCanonicalized<T> {
 /// outside this module -- converts from "canonical" universes to the
 /// original (but see the `UMapToCanonical` folder).
 #[derive(Clone, Debug)]
-pub(in solve) struct UniverseMap {
+pub(in ir::solve) struct UniverseMap {
     /// A reverse map -- for each universe Ux that appears in
     /// `quantified`, the corresponding universe in the original was
     /// `universes[x]`.
